SUG: Make question about current field more broad.
Current wording of the question seems to limit the scope in a
way that is not necessarily intended.
This would probably be addressed by what items are included
in the dropdown list, but I wanted to suggest updating the
wording so that it doesn't seem to limit respondents to
scientists and engineers.